src/backend/pyproject.toml
[tool.poetry]
name = "ai-template-backend"
version = "0.1.0"
description = "Backend of AI template"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.4"}
pydantic-settings = "^2.6.1"
sqlmodel = "^0.0.22"
alembic = "^1.14.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
passlib = "^1.7.4"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
emails = "^0.6"
elastic-apm = "^6.23.0"
bcrypt = "^4.2.0"
pydantic = "^2.9.2"
psutil = "^6.1.0"
boto3 = "^1.36.12"
psycopg2 = "^2.9.10"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
